+++ components/auth/loginForm.tsx
@@ -146,7 +146,7 @@ export const LoginFormComponent = () => {
 							</PopoverTrigger>
 							<PopoverContent>
 								<div className="px-1 py-2">
-									<div className="text-xs text-primary font-bold">Comming soon</div>
+									<div className="text-xs text-primary font-bold">Coming soon!</div>
 									<div className="text-xs">We are working on it</div>
 								</div>
 							</PopoverContent>