+++ components/home/hero.tsx
@@ -1,4 +1,3 @@
-import heroSVG from '@/public/hero.svg';
 import { Button } from '@nextui-org/button';
 import { Link } from '@nextui-org/link';
 import dynamic from 'next/dynamic';
@@ -8,33 +7,15 @@ const Image = dynamic(() => import('next/image'), { ssr: false });
 export const HeroComponent = () => {
 	return (
 		<div className="flex flex-col lg:flex-row items-center justify-between gap-8 py-12 px-4 md:px-12 lg:max-w-4xl mx-auto">
-			<div className="w-full lg:w-1/2">
-				<Image
-					src={heroSVG}
-					alt="Hero image for Komunal App"
-					width={300}
-					height={300}
-					priority={true}  // Added to prioritize loading
-					className="w-full h-auto max-w-[500px] mx-auto"
-				/>
-			</div>
-			<div className="w-full lg:w-1/2 space-y-4 text-center lg:text-left lg:mr-[3rem]">
-				<h1 className="text-2xl md:text-3xl lg:text-4xl font-bold">
-					Find <span className="text-primary">Answers</span>,
-					<br />
-					Share <span className="text-primary">Knowledge</span>!
+			<div className="w-full space-y-4 text-center">
+				<h1 className="text-4xl lg:text-7xl font-black">
+					Ask Your Friends <br />
+					<span className="text-primary">Questions</span>
 				</h1>
-				<p className="text-base md:text-lg py-1 w-full text-center lg:text-left">
-					A platform to ask and get answers from various topics.
-					Explore new knowledge every day.{' '}
+				<p className="text-xl lg:text-2xl py-1 w-[90%] text-center mx-auto">
+					Connect with friends, ask questions, and get real answers.
 				</p>
-				<Button
-					className="mt-4 text-sm md:text-base lg:text-lg px-4 py-2 md:px-6 md:py-3"
-					color="primary"
-					variant="solid"
-					as={Link}
-					href='/signup'
-				>
+				<Button className='mt-4 text-md  lg:text-lg px-4 py-2 md:px-6 md:py-3' color='primary' variant='solid' as={Link} href='/signup'>
 					Join Now
 				</Button>
 			</div>