+++ config/copywriting.ts
@@ -1,23 +1,25 @@
-
 export const benefitsCopy = [
 	{
-		index:1,
+		index: 1,
 		icon: "UsersFour",
 		header: "Open Community",
 		body: "Join a welcoming community always ready to help answer your questions.",
 	},
 	{
-		index:2,
+		index: 2,
+		icon: "BookOpen",
 		header: "Diverse Topics",
 		body: "From personal queries to hobbies, discuss a wide range of topics here.",
 	},
 	{
-		index:3,
+		index: 3,
+		icon: "ShieldCheck",
 		header: "Privacy Ensured",
 		body: "Ask questions with confidence, knowing your identity is protected.",
 	},
 	{
-		index:4,
+		index: 4,
+		icon: "ThumbsUp",
 		header: "Top Answer Ratings",
 		body: "Upvote the best answers to highlight quality content.",
 	},